refactor: move build env
