+++ crates/rattler_installs_packages/src/wheel_builder/mod.rs
@@ -1,236 +1,26 @@
 //! Turn an sdist into a wheel by creating a virtualenv and building the sdist in it
+
+mod build_environment;
+
 use parking_lot::Mutex;
 use std::sync::Arc;
-use std::{
-    collections::{HashMap, HashSet},
-    path::PathBuf,
-    process::{Command, Output},
-    str::FromStr,
-};
+use std::{collections::HashMap, path::PathBuf};
 
 use pep508_rs::{MarkerEnvironment, Requirement};
 
 use crate::{
     artifacts::wheel::UnpackError,
-    artifacts::wheel::UnpackWheelOptions,
     artifacts::SDist,
     artifacts::Wheel,
     index::PackageDb,
     python_env::WheelTags,
-    python_env::{PythonLocation, VEnv},
     types::Artifact,
     types::SDistFilename,
     types::{WheelCoreMetaDataError, WheelCoreMetadata},
 };
 
-use crate::resolve::{resolve, PinnedPackage, ResolveOptions, SDistResolution};
-
-// include static build_frontend.py string
-const BUILD_FRONTEND_PY: &str = include_str!("./wheel_builder_frontend.py");
-/// A build environment for building wheels
-/// This struct contains the virtualenv and everything that is needed
-/// to execute the PEP517 build backend hools
-#[derive(Debug)]
-pub struct BuildEnvironment<'db> {
-    work_dir: tempfile::TempDir,
-    package_dir: PathBuf,
-    #[allow(dead_code)]
-    build_system: pyproject_toml::BuildSystem,
-    entry_point: String,
-    build_requirements: Vec<Requirement>,
-    resolved_wheels: Vec<PinnedPackage<'db>>,
-    venv: VEnv,
-}
-
-impl<'db> BuildEnvironment<'db> {
-    /// Extract the wheel and write the build_frontend.py to the work folder
-    pub fn install_build_files(&self, sdist: &SDist) -> std::io::Result<()> {
-        // Extract the sdist to the work folder
-        sdist.extract_to(self.work_dir.path())?;
-        // Write the python frontend to the work folder
-        std::fs::write(
-            self.work_dir.path().join("build_frontend.py"),
-            BUILD_FRONTEND_PY,
-        )
-    }
-
-    /// Get the extra requirements and combine these to the existing requirements
-    /// This uses the `GetRequiresForBuildWheel` entry point of the build backend.
-    /// this might not be available for all build backends.
-    /// and it can also return an empty list of requirements.
-    fn get_extra_requirements(&self) -> Result<HashSet<Requirement>, WheelBuildError> {
-        let output = self.run_command("GetRequiresForBuildWheel").map_err(|e| {
-            WheelBuildError::CouldNotRunCommand("GetRequiresForBuildWheel".into(), e)
-        })?;
-
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            return Err(WheelBuildError::Error(stderr.to_string()));
-        }
-
-        // The extra requirements are stored in a file called extra_requirements.json
-        let extra_requirements_json =
-            std::fs::read_to_string(self.work_dir.path().join("extra_requirements.json"))?;
-        let extra_requirements: Vec<String> = serde_json::from_str(&extra_requirements_json)?;
-
-        Ok(HashSet::<Requirement>::from_iter(
-            extra_requirements
-                .iter()
-                .map(|s| Requirement::from_str(s).expect("...")),
-        ))
-    }
-
-    /// Install extra requirements into the venv, if any extra were found
-    /// If the extra requirements are already installed, this will do nothing
-    /// for that requirement.
-    async fn install_extra_requirements(
-        &self,
-        package_db: &'db PackageDb,
-        env_markers: &MarkerEnvironment,
-        wheel_tags: Option<&WheelTags>,
-        resolve_options: &ResolveOptions,
-    ) -> Result<(), WheelBuildError> {
-        // Get extra requirements if any
-        let extra_requirements = self.get_extra_requirements()?;
-
-        // Combine previous requirements with extra requirements
-        let combined_requirements = HashSet::from_iter(self.build_requirements.iter().cloned())
-            .union(&extra_requirements)
-            .cloned()
-            .collect::<Vec<_>>();
-
-        // Install extra requirements if any new ones were foujnd
-        if !extra_requirements.is_empty()
-            && self.build_requirements.len() != combined_requirements.len()
-        {
-            let locked_packages = HashMap::default();
-            // Todo: use the previous resolve for the favored packages?
-            let favored_packages = HashMap::default();
-            let all_requirements = combined_requirements.to_vec();
-            let extra_resolved_wheels = resolve(
-                package_db,
-                all_requirements.iter(),
-                env_markers,
-                wheel_tags,
-                locked_packages,
-                favored_packages,
-                resolve_options,
-            )
-            .await
-            .map_err(|_| WheelBuildError::CouldNotResolveEnvironment(all_requirements))?;
-
-            // install extra wheels
-            for package_info in extra_resolved_wheels {
-                if self.resolved_wheels.contains(&package_info) {
-                    continue;
-                }
-                tracing::info!(
-                    "installing extra requirements: {} - {}",
-                    package_info.name,
-                    package_info.version
-                );
-                let artifact_info = package_info.artifacts.first().unwrap();
-                let artifact = package_db
-                    .get_artifact::<Wheel>(artifact_info)
-                    .await
-                    .expect("could not get artifact");
-
-                self.venv
-                    .install_wheel(&artifact, &UnpackWheelOptions::default())?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Run a command in the build environment
-    fn run_command(&self, stage: &str) -> std::io::Result<Output> {
-        // three args: cache.folder, goal
-        Command::new(self.venv.python_executable())
-            .current_dir(&self.package_dir)
-            .arg(self.work_dir.path().join("build_frontend.py"))
-            .arg(self.work_dir.path())
-            .arg(&self.entry_point)
-            .arg(stage)
-            .output()
-    }
-
-    /// Setup the build environment so that we can build a wheel from an sdist
-    async fn setup(
-        sdist: &SDist,
-        package_db: &'db PackageDb,
-        env_markers: &MarkerEnvironment,
-        wheel_tags: Option<&WheelTags>,
-        resolve_options: &ResolveOptions,
-    ) -> Result<BuildEnvironment<'db>, WheelBuildError> {
-        // Setup a work directory and a new env dir
-        let work_dir = tempfile::tempdir().unwrap();
-        let venv = VEnv::create(&work_dir.path().join("venv"), PythonLocation::System).unwrap();
-
-        // Find the build system
-        let build_system =
-            sdist
-                .read_build_info()
-                .unwrap_or_else(|_| pyproject_toml::BuildSystem {
-                    requires: Vec::new(),
-                    build_backend: None,
-                    backend_path: None,
-                });
-        // Find the build requirements
-        let build_requirements = build_requirements(&build_system);
-        // Resolve the build environment
-        let resolved_wheels = resolve(
-            package_db,
-            build_requirements.iter(),
-            env_markers,
-            wheel_tags,
-            HashMap::default(),
-            HashMap::default(),
-            resolve_options,
-        )
-        .await
-        .map_err(|_| WheelBuildError::CouldNotResolveEnvironment(build_requirements.to_vec()))?;
-
-        // Install into venv
-        for package_info in resolved_wheels.iter() {
-            let artifact_info = package_info.artifacts.first().unwrap();
-            let artifact = package_db
-                .get_artifact::<Wheel>(artifact_info)
-                .await
-                .map_err(|_| WheelBuildError::CouldNotGetArtifact)?;
-
-            venv.install_wheel(
-                &artifact,
-                &UnpackWheelOptions {
-                    installer: None,
-                    ..Default::default()
-                },
-            )?;
-        }
-
-        const DEFAULT_BUILD_BACKEND: &str = "setuptools.build_meta:__legacy__";
-        let entry_point = build_system
-            .build_backend
-            .clone()
-            .unwrap_or_else(|| DEFAULT_BUILD_BACKEND.to_string());
-
-        // Package dir for the package we need to build
-        let package_dir = work_dir.path().join(format!(
-            "{}-{}",
-            sdist.name().distribution.as_source_str(),
-            sdist.name().version
-        ));
-
-        Ok(BuildEnvironment {
-            work_dir,
-            package_dir,
-            build_system,
-            build_requirements,
-            entry_point,
-            resolved_wheels,
-            venv,
-        })
-    }
-}
+use crate::resolve::{ResolveOptions, SDistResolution};
+use crate::wheel_builder::build_environment::BuildEnvironment;
 
 type BuildCache<'db> = Mutex<HashMap<SDistFilename, Arc<BuildEnvironment<'db>>>>;
 
@@ -425,8 +215,7 @@ impl<'db, 'i> WheelBuilder<'db, 'i> {
             return Err(WheelBuildError::Error(stdout.to_string()));
         }
 
-        let result =
-            std::fs::read_to_string(build_environment.work_dir.path().join("metadata_result"))?;
+        let result = std::fs::read_to_string(build_environment.work_dir().join("metadata_result"))?;
         let folder = PathBuf::from(result.trim());
         let path = folder.join("METADATA");
 
@@ -448,8 +237,7 @@ impl<'db, 'i> WheelBuilder<'db, 'i> {
             return Err(WheelBuildError::Error(stdout.to_string()));
         }
 
-        let result =
-            std::fs::read_to_string(build_environment.work_dir.path().join("wheel_result"))?;
+        let result = std::fs::read_to_string(build_environment.work_dir().join("wheel_result"))?;
         let wheel_file = PathBuf::from(result.trim());
 
         Ok(wheel_file)
